Update tf_deploy_win_hosts.tf
Warning: Interpolation-only expressions are deprecated

  on tf_deploy_win_hosts.tf line 118, in resource "azurerm_virtual_machine" "myterraformvm":
 118:     resource_group_name   = "${azurerm_resource_group.myterraformgroup.name}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.